# src/openapi/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Axie Data Hub",
    "description": "Axie Data Hub is a collection of different APIs that provide information about Axie Infinity.",
    "version": "1.1.0",
    "contact": {
      "name": "Axie Data Hub",
      "url": "https://rapidapi.com/marsigliabusiness--MCJDzcrYSM/api/axie-data-hub",
      "email": "marsiglia.business@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1.1"
    },
    {
      "url": "https://axie-infinity-api.herokuapp.com/api/v1.1"
    },
    {
      "url": "https://axie-infinity-api.vercel.app/api/v1.1"
    },
    {
      "url": "https://axie-data-hub.glitch.me/api/v1.1"
    },
    {
      "url": "https://axie-infinity-api-production.up.railway.app/api/v1.1"
    },
    {
      "url": "https://axie-data-hub.p.rapidapi.com"
    }
  ],
  "paths": {
...
    "/game-data/classic/cards/{patchId}": {
      "get": {
        "summary": "Get classic ability cards data from specific patch.",
        "description": "Get classic ability cards data from specific patch. You can filter by any valid card object property, such as 'class' or 'type'. E.g. '/cards/classic/current?class=reptile&type=mouth'. Theese filters are not available if you use 'format=original'",
        "parameters": [
          {
            "name": "patchId",
            "in": "path",
            "description": "Patch ID.",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "current"
          },
          {
            "name": "format",
            "in": "query",
            "description": "Format of the response: 'default' or 'original'. Original has the original format response given by Axie Infinity. Default is a better formatted version of the original response. Default format is 'default'.",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "default"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/cardsData"
                }
              }
            }
          }
        }
      }
    },
    "/game-data/origin/cards": {
      "get": {
        "summary": "Get origin ability cards data.",
        "description": "Get origin ability cards data. You can filter by any valid card object property, such as 'class', 'type' or 'abilityType'. E.g. '/cards/origin?class=reptile&type=mouth&abilityType=AttackRanged'. Theese filters are not available if you use 'format=original'",
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "description": "Format of the response: 'default' or 'original'. Original has the original format response given by Axie Infinity. Default is a better formatted version of the original response. Default format is 'default'.",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "default"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/originCardsData"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "cardsData": {},
      "originCardsData": {},
...
    }
  }
}